refactored decorator and classbased
